Inspectro pip module; cis eigdecomp added.
