+++ inspectro/__init__.py
@@ -1,0 +1,10 @@
+"""
+inspectro
+~~~~~~~
+Spectral decomposition for characterizing long-range interaction profiles in Hi-C maps.
+:copyright: (c) 2017-2022 Open2C
+:author: Open2C
+:license: MIT
+"""
+
+__version__ = "0.0.1"
