+++ inspectro/__main__.py
@@ -1,0 +1,2 @@
+from .utils import clustering, eigdecomp, df2multivec, plotting, common
+
