+++ pyproject.toml
@@ -1,0 +1,61 @@
+[build-system]
+requires = ["cython",
+            "numpy>=1.10",
+            "setuptools >= 43.0.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+py-modules = ["inspectro"]
+
+[tool.black]
+line-length = 127
+target-version = ['py39', 'py310']
+
+
+[project.scripts]
+pairtools = "inspectro.utils:utils"
+
+[project]
+name = "inspectro"
+version = "0.0.1"
+license = {file = "LICENSE"}
+description = "Spectral decomposition for characterizing long-range interaction profiles in Hi-C maps"
+maintainers = [
+ {name = "Open2C", email = "open.chromosome.collective@gmail.com"},
+]
+requires-python = ">=3.9"
+dependencies = ["cython",
+            "numpy>=1.10",
+            "h5py",
+            "bioframe>=0.3.3",
+            "cooltools>=0.5.2",
+            "cooler",
+            "pandas",
+            "scipy",
+            "pybbi",
+            "tqdm",
+            "scikit-learn",
+            "seaborn",
+            "matplotlib>=3.4.0",
+            "datashader",
+            "ipympl",
+            "ipdb",
+            "jupyterlab",
+            "loky",
+            "pyarrow",
+            "pysam",
+            "snakemake"]
+
+readme = "README.md"
+classifiers = [
+ "Development Status :: 4 - Beta",
+ "Operating System :: OS Independent",
+ "Programming Language :: Python",
+ "Programming Language :: Python :: 3",
+ "Programming Language :: Python :: 3.9",
+ "Programming Language :: Python :: 3.10",
+]
+
+[project.urls]
+homepage = "https://github.com/open2c/inspectro"
+source = "https://github.com/open2c/inspectro"
